src/requirements.txt
#Core dependencies
psycopg2-binary>=2.9.7
python-dotenv>=1.0.1
boto3>=1.26.0
pandas>=2.1.2
pyarrow
s3fs
requests

